Refactoring the code and adding serverless
