+++ .gitignore
@@ -1,0 +1,21 @@
+# Distribution / packaging
+.Python
+env/
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+*.egg-info/
+.installed.cfg
+*.egg
+*.__pycache__
+.pytest_cache
+# Serverless directories
+.serverless
