+++ requirements-lambda.txt
@@ -1,0 +1,2 @@
+opensearch-py
+user-agents
