+++ requirements.txt
@@ -1,0 +1,4 @@
+boto3
+opensearch-py
+user-agents
+pytest
